Updated imports after restructuring for python package creation
