+++ .github/workflows/pytest.yml
@@ -2,15 +2,15 @@ name: pytest
 
 on:
   push:
-    branches: [ main ]
+    branches: [main]
   pull_request:
-    branches: [ main ]
+    branches: [main]
 
 jobs:
   pyTestCov:
     strategy:
       matrix:
-        python-version: ['3.10', '3.11']
+        python-version: ["3.10", "3.11"]
     runs-on: ubuntu-latest
     steps:
       - name: Python Setup
@@ -26,6 +26,7 @@ jobs:
           pip install -r requirements.txt
           pip install pytest-timeout
           pip install pytest-cov
+          pip install .
       - name: Run Tests
         run: python -m pytest -v --timeout=60
       - name: Upload to Codecov
