+++ src/esoh/ingest/main.py
@@ -1,5 +1,5 @@
-from ingest.send_mqtt import mqtt_connection
-from ingest.messages import load_files, build_message
+from esoh.ingest.send_mqtt import mqtt_connection
+from esoh.ingest.messages import load_files, build_message
 
 from pathlib import Path
